Split notes-api.spec.js into per-entity test files for Stage 2 lane ownership
Pure refactoring of test organization -- no tests added or removed (85 total).

Split the monolithic notes-api.spec.js into entity-scoped files so that
Stage 2 entity lanes (comments, study-notes, reports) can each own their
own tests without merge conflicts:

- comments.spec.js: suites 28-30 (add, edit, delete comments)
- study-notes.spec.js: suites 26-27 (study/series descriptions)
- reports.spec.js: suites 31-33, 35 (upload, download, delete, ID validation)
- notes-api.spec.js: suites 25, 34, 36 (batch load, migration, lifecycle)
- notes-test-helpers.js: shared helpers (uniqueStudyUid, uploadReport, etc.)

desktop-report-persistence.spec.js reviewed and already scoped to reports
only (2 tests) -- no changes needed.

mock-tauri-sql-init.js does not exist; each desktop test file has its own
inline mock tailored to its needs -- no consolidation needed.
